Refactor RouterConfig into separate file
